derp

+++ labels/kmeans2.py
@@ -28,7 +28,7 @@ for h5fn in config.h5s[2:]:
             continue
         image, _ = datam.handle_images([(h5f, df)])
         print(image.shape)
-        pred = model.predict(image).reshape(1024,1024)
+        pred = model.predict(image).astype(np.uint8).reshape(1024,1024)
         fname = "%s/kmeans_labels/%s_%s_kmeans_anno.png" % (
             config.data_path, h5fn[:1], df)
         imwrite(fname, pred)
